+++ app/src/Camera.hs
@@ -5,9 +5,9 @@ module Camera where
 import Vector
 import Matrix
 
-data Camera = Camera { worldToCamera :: Matrix, fieldOfView :: !Float, position :: Vector } deriving (Show)
+data Camera = Camera { worldToCamera :: Matrix, fieldOfView :: !Double, position :: Vector } deriving (Show)
 
-lookAt :: Position -> Position -> Direction -> Float -> Camera
+lookAt :: Position -> Position -> Direction -> Double -> Camera
 lookAt pos target up fov = 
     Camera matrix fov pos
     where
@@ -16,7 +16,7 @@ lookAt pos target up fov =
       up' = right `cross` forward
       matrix = buildMatrix right up' forward pos
 
-withVectors :: Position -> Direction -> Direction -> Direction -> Float -> Camera
+withVectors :: Position -> Direction -> Direction -> Direction -> Double -> Camera
 withVectors pos basisX basisY basisZ fov = Camera matrix fov pos
     where
       matrix = buildMatrix basisX basisY basisZ (Vector.negate pos)
